make refactoring for all pages. Extend all-pages from list-page component. Remove btn new test from All test page heading. Add scss file for mixins
