+++ src/app/app.module.ts
@@ -11,17 +11,13 @@ import { AllRespondentsComponent } from './pages/all-respondents/all-respondents
 import { HeaderComponent } from './shared/components/header/header.component';
 import { NavigationButtonsComponent } from './shared/components/header/navigation-buttons/navigation-buttons.component';
 import { ToolsComponent } from './shared/components/header/tools/tools.component';
-import { WraperComponent } from './pages/all-tests/tests-list/tests-list.component';
 import { CreateNewBlockComponent } from './pages/all-tests/create-new-block/create-new-block.component';
-import { TestItemComponent } from './pages/all-tests/tests-list/test-item/test-item.component';
+import { TestItemComponent } from './pages/all-tests/test-item/test-item.component';
 import { CreateNewTestComponent } from './pages/create-new-test/create-new-test.component';
 import { CreateNewTestUploaderComponent } from './pages/create-new-test/create-new-test-uploader/create-new-test-uploader.component';
-import { AllRespondentsListComponent } from './pages/all-respondents/all-respondents-list/all-respondents-list.component';
-import { RespondentItemComponent } from './pages/all-respondents/all-respondents-list/respondent-item/respondent-item.component';
+import { RespondentItemComponent } from './pages/all-respondents/respondent-item/respondent-item.component';
 import { CreateNewComponent } from './shared/components/create-new/create-new.component';
-import { ActivationsListWrapperComponent } from './pages/all-activations/activations-list-wrapper/activations-list-wrapper.component';
-import { ActivationItemComponent } from './pages/all-activations/activations-list-wrapper/activation-item/activation-item.component';
-import { PageInformationComponent } from './pages/all-activations/page-information/page-information.component';
+import { ActivationItemComponent } from './pages/all-activations/activation-item/activation-item.component';
 import { HorizontalLineComponent } from './shared/components/horizontal-line/horizontal-line.component';
 import { RespondentService} from './shared/services/respondent.service';
 import { SelectComponent } from './shared/components/select/select.component';
@@ -31,6 +27,7 @@ import { FileuploaderComponent } from './pages/create-new-test/question-list/fil
 import { TestService } from './shared/services/test.service';
 import { HttpClientModule } from '@angular/common/http';
 import { ModalComponent } from './shared/components/modal/modal.component';
+import { ListPageComponent } from './pages/layout/list-page/list-page.component';
 
 
 @NgModule({
@@ -42,25 +39,22 @@ import { ModalComponent } from './shared/components/modal/modal.component';
     HeaderComponent,
     NavigationButtonsComponent,
     ToolsComponent,
-    WraperComponent,
     CreateNewBlockComponent,
     TestItemComponent,
     CreateNewTestComponent,
     CreateNewTestUploaderComponent,
-    AllRespondentsListComponent,
     RespondentItemComponent,
     TestItemComponent,
     CreateNewComponent,
-    ActivationsListWrapperComponent,
     ActivationItemComponent,
-    PageInformationComponent,
     HorizontalLineComponent,
     CreateNewComponent,
     SelectComponent,
     FileuploaderComponent,
     QuestionFormComponent,
     QuestionComponent,
-    ModalComponent
+    ModalComponent,
+    ListPageComponent
   ],
   imports: [
     BrowserModule,
