+++ src/app/pages/all-activations/activation-item/activation-item.component.html
