--- src/app/pages/all-activations/activations-list-wrapper/activations-list-wrapper.component.html
@@ -1,3 +1,0 @@
-<div class="all-activations-wrapper">
-  <pt-activation-item *ngFor="let activation of activations" [activation]="activation"></pt-activation-item>
-</div>
