--- src/app/pages/all-activations/activations-list-wrapper/activations-list-wrapper.component.scss
@@ -1,6 +1,0 @@
-.all-activations-wrapper {
-  margin-top: 48px;
-  max-width: 999px;
-  margin-left: auto;
-  margin-right: auto;
-}
