+++ src/app/pages/all-activations/all-activations.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Activation } from '../models/activation';
 
 @Component({
   selector: 'pt-all-activations',
@@ -6,6 +7,11 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./all-activations.component.scss']
 })
 export class AllActivationsComponent implements OnInit {
+  activations: Activation[] = [
+    new Activation('ACTIVATION NAME', 2, 1, 3, '/assets/images/image.jpeg', true),
+    new Activation('ANOTHER ACTIVATOIN', 4, 5, 6, '/assets/images/image.jpeg', false),
+  ];
+
   constructor() { }
 
   ngOnInit() {
