--- src/app/pages/all-activations/page-information/page-information.component.html
@@ -1,4 +1,0 @@
-<div class="page-information">
-    <span class="activations-left">ALL ACTIVATIONS</span> <span class="activations-right">all activations</span>
-</div>
-<pt-horizontal-line></pt-horizontal-line>
