--- src/app/pages/all-activations/page-information/page-information.component.scss
@@ -1,28 +1,0 @@
-@import "src/assets/styles/colors";
-
-.page-information {
-  margin-top: 140px;
-  width: 999px;
-  margin-left: auto;
-  margin-right: auto;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-  justify-content: space-between;
-
-}
-
-span {
-  font-family: "Roboto", sans-serif;
-  color: $white;
-  font-weight: 200;
-
-}
-
-.activations-left {
-  font-size: 34px;
-}
-
-.activations-right {
-  font-size: 20px;
-}
