--- src/app/pages/all-activations/page-information/page-information.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'pt-page-information',
-  templateUrl: './page-information.component.html',
-  styleUrls: ['./page-information.component.scss']
-})
-export class PageInformationComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
