--- src/app/pages/all-respondents/all-respondents-list/all-respondents-list.component.html
@@ -1,15 +1,0 @@
-<div class="testTitle">
-	<div class="title-left">
-		<div class="title-for-list">All Respondents <span class="title-for-list__quantity">(2)</span></div>
-	</div>
-	<div class="title-right">
-		<div class="close-menu">All Respondents</div>
-		<div class="angle-down-respondents">
-			<i class="fa fa-angle-down"></i>
-		</div>
-	</div>
-</div>
-<div class="respondent__items">
-	<pt-respondent-item *ngFor="let respondent of respondentsList" [respondent]="respondent"></pt-respondent-item>
-</div>
-
