--- src/app/pages/all-respondents/all-respondents-list/all-respondents-list.component.scss
@@ -1,51 +1,0 @@
-@import 'src/assets/styles/colors';
-
-$width: 1440px;
-$font-size: 20px;
-
-.testTitle {
-  display: flex;
-  justify-content: space-between;
-  padding-top: 135px;
-  padding-bottom: 15px;
-  border-bottom: solid 2px rgba(255, 255, 255, 0.41);
-}
-
-.title-for-list {
-  text-transform: uppercase;
-  font-size: 34px;
-  color: #ffffff;
-
-  &__quantity {
-    padding-bottom: 15px;
-    border-bottom: 2px solid #fff;
-  }
-}
-
-.title-right {
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  overflow: auto;
-  color: $white;
-}
-
-.close-menu {
-  font-size: 20px;
-  font-weight: 300;
-}
-
-.angle-down-respondents {
-  margin-left: 10px;
-}
-
-.fa-angle-down {
-  font-size: 26px;
-}
-
-.respondent {
-  &__items {
-    margin-top: 35px;
-  }
-}
