--- src/app/pages/all-respondents/all-respondents-list/all-respondents-list.component.ts
@@ -1,21 +1,0 @@
-import {Component, OnInit} from '@angular/core';
-import {RespondentModel} from '../../models/respondent.model';
-import {RespondentService} from '../../../shared/services/respondent.service';
-
-@Component({
-  selector: 'pt-all-respondents-list',
-  templateUrl: './all-respondents-list.component.html',
-  styleUrls: ['./all-respondents-list.component.scss']
-})
-export class AllRespondentsListComponent implements OnInit {
-
-  public respondentsList: RespondentModel[];
-
-  constructor(private respondentService: RespondentService) {
-  }
-
-  ngOnInit() {
-    this.respondentService.getRespondents().subscribe(result => this.respondentsList = result);
-  }
-
-}
