+++ src/app/pages/all-respondents/all-respondents.component.html
@@ -1,6 +1,4 @@
 <pt-header></pt-header>
-<div class="app">
-	<div class="app__container">
-		<pt-all-respondents-list></pt-all-respondents-list>
-	</div>
-</div>
+<pt-list-page heading="All Respondents" listLength="{{respondentsList.length}}">
+	<pt-respondent-item *ngFor="let respondent of respondentsList" [respondent]="respondent"></pt-respondent-item>
+</pt-list-page>
