+++ src/app/pages/all-respondents/respondent-item/respondent-item.component.html
@@ -1,5 +1,4 @@
-<div class="wrapper">
-  <div class="test-item">
+<div class="test-item">
     <div class="photo"></div>
     <div class="description-summary">
       <div class="last-visited">Joined 5 minutes ago</div>
@@ -32,4 +31,3 @@
       </div>
     </div>
   </div>
-</div>
