+++ src/app/pages/all-respondents/respondent-item/respondent-item.component.scss
