+++ src/app/pages/all-tests/test-item/test-item.component.ts
@@ -1,5 +1,5 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {Test} from '../../../models/test';
+import {Test} from '../../models/test';
 
 @Component({
   selector: 'pt-test-item',
