--- src/app/pages/all-tests/tests-list/tests-list.component.html
@@ -1,3 +1,0 @@
-<div>
-    <pt-test-item *ngFor="let test of tests" [test]="test"></pt-test-item>
-</div>
