--- src/app/pages/all-tests/tests-list/tests-list.component.scss
@@ -1,4 +1,0 @@
-div {
-     width: 1000px;
-    margin: 0 auto;
-}
