+++ src/app/pages/layout/list-page/list-page.component.scss
@@ -1,0 +1,57 @@
+@import 'src/assets/styles/colors';
+@import 'src/assets/styles/mixins';
+
+
+.list-page {
+  margin-top: 75px;
+  &__wrapper {
+    max-width: 1000px;
+    margin: 0 auto;
+  }
+  &__header {
+    display: flex;
+    justify-content: space-between;
+    align-content: center;
+    border-bottom: 2px solid rgba(255, 255, 255, 0.41);
+    margin-bottom: 50px;
+  }
+  &__heading {
+    font-size: 34px;
+    font-weight: 400;
+    text-transform: uppercase;
+    color: $white;
+    margin: 0;
+    padding-bottom: 15px;
+  }
+  &__length {
+    position: relative;
+    margin: 0 10px;
+    &:after {
+      content: '';
+      width: 100%;
+      height: 2px;
+      background: $white;
+      position: absolute;
+      left: 0;
+      bottom: -17px; // heading padding-bottom + this.height
+    }
+  }
+  &__rpart {
+    align-self: center;
+  }
+  &__select {
+    font-size: 20px;
+    font-weight: 200;
+    color: $white;
+  }
+}
+
+.pt-select {
+  position: relative;
+  &__item {
+    @extend %reset-browser-ui;
+  }
+  &:after {
+    content: '';
+  }
+}
