+++ src/styles.scss
@@ -1,6 +1,7 @@
 /* You can add global styles to this file, and also import other style files */
 
 @import 'src/assets/styles/normalize';
+@import 'src/assets/styles/mixins';
 @import 'src/assets/styles/colors';
 @import 'src/assets/styles/fonts';
